Move AgreementResponseHelper to be with controllers
Since the Agreements Controller will be using it
